Split  OpenAIModelProvider
