+++ AFAAS/configs/config.py
@@ -6,14 +6,10 @@ import re
 from pathlib import Path
 from typing import Any, Optional, Union
 
-import autogpt
 import yaml
 from auto_gpt_plugin_template import AutoGPTPluginTemplate
 from AFAAS.configs.schema  import Configurable, SystemSettings, UserConfigurable
-from AFAAS.core.adapters.openai.openai import OPEN_AI_CHAT_MODELS, OpenAICredentials
-from autogpt.logs.config import LogFormatName, LoggingConfig
-from autogpt.plugins.plugins_config import PluginsConfig
-from autogpt.speech import TTSConfig
+from AFAAS.core.adapters.openai.chatmodel import OPEN_AI_CHAT_MODELS, OpenAICredentials
 from colorama import Fore
 from pydantic import Field, validator
