+++ AFAAS/core/adapters/openai/__init__.py
@@ -1,1 +1,1 @@
-from .openai import *
+from .chatmodel import *
