+++ AFAAS/core/adapters/openai/chatmodel.py
@@ -1,23 +1,23 @@
 import enum
-import functools
 import math
 import os
-import time
+import tiktoken
 from typing import Any, Callable, Dict, ParamSpec, Tuple, TypeVar, ClassVar
 
-from openai import AsyncOpenAI, APIError, RateLimitError, completions
+from openai import AsyncOpenAI, completions
 from openai.resources import AsyncCompletions, AsyncEmbeddings
 
-aclient = AsyncOpenAI(api_key=os.environ["OPENAI_API_KEY"])
-import tiktoken
+from AFAAS.core.adapters.openai.common import _OpenAIRetryHandler, OPEN_AI_CHAT_MODELS, OPEN_AI_EMBEDDING_MODELS, OPEN_AI_MODELS, OpenAIModelName, OpenAISettings, OpenAIConfiguration, OpenAIModelProviderBudget
+from AFAAS.core.adapters.openai.embeddings import _create_embedding  
 
-from AFAAS.lib.sdk.logger import AFAASLogger
+aclient = AsyncOpenAI(api_key=os.environ["OPENAI_API_KEY"])
+
 from AFAAS.configs import (Configurable,
                                                          SystemConfiguration,
                                                          UserConfigurable)
 from AFAAS.interfaces.adapters.chat_model import (AbstractChatMessage,
-    AssistantChatMessageDict, AssistantToolCallDict, BaseChatModelProvider,
-    AbstractRoleLabels, ChatMessage, ChatModelInfo, ChatModelResponse, CompletionModelFunction)
+    AssistantChatMessageDict, BaseChatModelProvider,
+    AbstractRoleLabels, ChatMessage, ChatModelResponse, CompletionModelFunction)
 from AFAAS.interfaces.adapters.language_model import (
     BaseModelProviderBudget, BaseModelProviderConfiguration, BaseModelProviderCredentials,
     BaseModelProviderSettings, BaseModelProviderUsage, Embedding,
@@ -25,6 +25,7 @@ from AFAAS.interfaces.adapters.language_model import (
     ModelProviderName, ModelProviderService, ModelTokenizer)
 from AFAAS.lib.utils.json_schema import JSONSchema
 
+from AFAAS.lib.sdk.logger import AFAASLogger
 LOG = AFAASLogger(name=__name__)
 
 _T = TypeVar("_T")
@@ -41,154 +42,13 @@ class OpenAIRoleLabel(AbstractRoleLabels):
 
     FUNCTION = "function"
     """May be used for the return value of function calls"""
+
 class OpenAIChatMessage(AbstractChatMessage):
     _role_labels: ClassVar[OpenAIRoleLabel] = OpenAIRoleLabel()
 
-class OpenAIModelName(str, enum.Enum):
-    """
-    Enumeration of OpenAI model names.
 
-    Attributes:
-        Each enumeration represents a distinct OpenAI model name.
-    """
-
-    # ADA = "text-embedding-ada-002"
-    # GPT3 = "gpt-3.5-turbo-instruct"
-    # GPT3_16k = "gpt-3.5-turbo-1106"
-    # GPT3_FINE_TUNED = "gpt-3.5-turbo-1106" + ""
-    # # GPT4 = "gpt-4-0613" # TODO for tests
-    # GPT4 = "gpt-3.5-turbo-1106"
-    # GPT4_32k = "gpt-4-1106-preview"
-
-    ADA = "text-embedding-ada-002"
-    GPT3 = "gpt-3.5-turbo-1106"
-    GPT3_16k = "gpt-3.5-turbo-1106"
-    GPT3_FINE_TUNED = "gpt-3.5-turbo-1106" + ""
-    # GPT4 = "gpt-3.5-turbo-1106" # TODO for tests
-    GPT4 = "gpt-3.5-turbo-1106"
-    GPT4_32k = "gpt-3.5-turbo-1106"
-
-
-OPEN_AI_EMBEDDING_MODELS = {
-    OpenAIModelName.ADA: EmbeddingModelInfo(
-        name=OpenAIModelName.ADA,
-        service=ModelProviderService.EMBEDDING,
-        provider_name=ModelProviderName.OPENAI,
-        prompt_token_cost=0.0001 / 1000,
-        max_tokens=8191,
-        embedding_dimensions=1536,
-    ),
-}
-
-
-OPEN_AI_CHAT_MODELS = {
-    info.name: info
-    for info in [
-        ChatModelInfo(
-            name=OpenAIModelName.GPT3,
-            service=ModelProviderService.CHAT,
-            provider_name=ModelProviderName.OPENAI,
-            prompt_token_cost=0.0015 / 1000,
-            completion_token_cost=0.002 / 1000,
-            max_tokens=4096,
-            has_function_call_api=True,
-        ),
-        ChatModelInfo(
-            name=OpenAIModelName.GPT3_16k,
-            service=ModelProviderService.CHAT,
-            provider_name=ModelProviderName.OPENAI,
-            prompt_token_cost=0.003 / 1000,
-            completion_token_cost=0.004 / 1000,
-            max_tokens=16384,
-            has_function_call_api=True,
-        ),
-        ChatModelInfo(
-            name=OpenAIModelName.GPT3_FINE_TUNED,
-            service=ModelProviderService.CHAT,
-            provider_name=ModelProviderName.OPENAI,
-            prompt_token_cost=0.0120 / 1000,
-            completion_token_cost=0.0160 / 1000,
-            max_tokens=4096,
-            has_function_call_api=True,
-        ),
-        ChatModelInfo(
-            name=OpenAIModelName.GPT4,
-            service=ModelProviderService.CHAT,
-            provider_name=ModelProviderName.OPENAI,
-            prompt_token_cost=0.03 / 1000,
-            completion_token_cost=0.06 / 1000,
-            max_tokens=8191,
-            has_function_call_api=True,
-        ),
-        ChatModelInfo(
-            name=OpenAIModelName.GPT4_32k,
-            service=ModelProviderService.CHAT,
-            provider_name=ModelProviderName.OPENAI,
-            prompt_token_cost=0.06 / 1000,
-            completion_token_cost=0.12 / 1000,
-            max_tokens=32768,
-            has_function_call_api=True,
-        ),
-    ]
-}
-
-
-OPEN_AI_MODELS = {
-    **OPEN_AI_CHAT_MODELS,
-    **OPEN_AI_EMBEDDING_MODELS,
-}
-
-
-class OpenAIConfiguration(BaseModelProviderConfiguration):
-    """Configuration for OpenAI.
-
-    Attributes:
-        retries_per_request: The number of retries per request.
-        maximum_retry: The maximum number of retries allowed.
-        maximum_retry_before_default_function: The maximum number of retries before a default function is used.
-    """
-
-    retries_per_request: int = UserConfigurable(default=10)
-    maximum_retry: int = 1
-    maximum_retry_before_default_function: int = 1
-
-
-class OpenAIModelProviderBudget(BaseModelProviderBudget):
-    """Budget configuration for the OpenAI Model Provider.
-
-    Attributes:
-        graceful_shutdown_threshold: The threshold for graceful shutdown.
-        warning_threshold: The warning threshold for budget.
-    """
-
-    graceful_shutdown_threshold: float = UserConfigurable(default=0.005)
-    warning_threshold: float = UserConfigurable(default=0.01)
-
-    total_budget: float = math.inf
-    total_cost: float = 0.0
-    remaining_budget: float = math.inf
-    usage: BaseModelProviderUsage = BaseModelProviderUsage()
-
-
-class OpenAISettings(BaseModelProviderSettings):
-    """Settings for the OpenAI provider.
-
-    Attributes:
-        configuration: Configuration settings for OpenAI.
-        credentials: The credentials for the model provider.
-        budget: Budget settings for the model provider.
-    """
-
-    configuration: OpenAIConfiguration = OpenAIConfiguration()
-    credentials: BaseModelProviderCredentials = BaseModelProviderCredentials()
-    budget: OpenAIModelProviderBudget = OpenAIModelProviderBudget()
-
-    name = "chat_model_provider"
-    description = "Provides access to OpenAI's API."
-
-
-class OpenAIProvider(
-    Configurable[OpenAISettings], BaseChatModelProvider, EmbeddingModelProvider
+class AFAASChatOpenAI(
+    Configurable[OpenAISettings], BaseChatModelProvider
 ):
     """A provider for OpenAI's API.
 
@@ -199,6 +59,7 @@ class OpenAIProvider(
     """
 
     default_settings = OpenAISettings()
+    chatas = OpenAIChatMessage
 
     def __init__(
         self,
@@ -567,44 +428,6 @@ class OpenAIProvider(
         )
         return await self.create_chat_completion(**kwargs)
 
-    async def create_embedding(
-        self,
-        text: str,
-        model_name: OpenAIModelName,
-        embedding_parser: Callable[[Embedding], Embedding],
-        **kwargs,
-    ) -> EmbeddingModelResponse:
-        """Create an embedding using the OpenAI API.
-
-        Args:
-            text (str): The text to embed.
-            model_name (str): The name of the embedding model.
-            embedding_parser (Callable): A parser to process the embedding.
-            **kwargs: Additional keyword arguments.
-
-        Returns:
-            EmbeddingModelResponse: Response containing the embedding.
-
-        Example:
-            >>> provider = OpenAIProvider(...)
-            >>> embedding_response = await provider.create_embedding("Hello, world!", ...)
-            >>> print(embedding_response.embedding)
-            [0.123, -0.456, ...]
-        """
-        embedding_kwargs = self._get_embedding_kwargs(model_name, **kwargs)
-        response = await self._create_embedding(text=text, **embedding_kwargs)
-
-        response_args = {
-            "model_info": OPEN_AI_EMBEDDING_MODELS[model_name],
-            "prompt_tokens_used": response.usage.prompt_tokens,
-            "completion_tokens_used": response.usage.completion_tokens,
-        }
-        response = EmbeddingModelResponse(
-            **response_args,
-            embedding=embedding_parser(response.embeddings[0]),
-        )
-        self._budget.update_usage_and_cost(response)
-        return response
 
     def _get_completion_kwargs(
         self,
@@ -637,39 +460,9 @@ class OpenAIProvider(
             completion_kwargs["tools"] = [
                 {"type": "function", "function": f.schema} for f in functions
             ]
-        else:
-            # Provide compatibility with older models
-            _functions_compat_fix_kwargs(functions, completion_kwargs)
 
         return completion_kwargs
 
-    def _get_embedding_kwargs(
-        self,
-        model_name: OpenAIModelName,
-        **kwargs,
-    ) -> dict:
-        """Get kwargs for embedding API call.
-
-        Args:
-            model: The model to use.
-            kwargs: Keyword arguments to override the default values.
-
-        Returns:
-            dict: Dictionary containing the kwargs.
-
-        Example:
-            >>> provider = OpenAIProvider(...)
-            >>> embedding_kwargs = provider._get_embedding_kwargs(OpenAIModelName.ADA, ...)
-            >>> print(embedding_kwargs)
-            {'model': 'text-embedding-ada-002', ...}
-        """
-        embedding_kwargs = {
-            "model": model_name,
-            **kwargs,
-            **self._credentials.unmasked(),
-        }
-
-        return embedding_kwargs
 
     def __repr__(self):
         """
@@ -688,19 +481,6 @@ class OpenAIProvider(
     def has_oa_tool_calls_api(self, model_name: str) -> bool:
         # print(self._providers[model_name])
         return OPEN_AI_CHAT_MODELS[model_name].has_function_call_api
-
-
-async def _create_embedding(text: str, *_, **kwargs) -> AsyncEmbeddings:
-    """Embed text using the OpenAI API.
-
-    Args:
-        text str: The text to embed.
-        model_name str: The name of the model to use.
-
-    Returns:
-        str: The embedding.
-    """
-    return await aclient.embeddings.create(input=[text], **kwargs)
 
 
 async def _create_chat_completion(
@@ -740,205 +520,3 @@ async def _create_chat_completion(
         messages=raw_messages, **kwargs
     )
     return return_value
-
-
-class _OpenAIRetryHandler:
-    """Retry Handler for OpenAI API call.
-
-    Args:
-        num_retries int: Number of retries. Defaults to 10.
-        backoff_base float: Base for exponential backoff. Defaults to 2.
-        warn_user bool: Whether to warn the user. Defaults to True.
-    """
-
-    _retry_limit_msg = "Error: Reached rate limit, passing..."
-    _api_key_error_msg = (
-        "Please double check that you have setup a PAID OpenAI API Account. You can "
-        "read more here: https://docs.agpt.co/setup/#getting-an-api-key"
-    )
-    _backoff_msg = "Error: API Bad gateway. Waiting {backoff} seconds..."
-
-    def __init__(
-        self,
-        
-        num_retries: int = 10,
-        backoff_base: float = 2.0,
-        warn_user: bool = True,
-    ):
-        self._num_retries = num_retries
-        self._backoff_base = backoff_base
-        self._warn_user = warn_user
-
-    def _log_rate_limit_error(self) -> None:
-        LOG.trace(self._retry_limit_msg)
-        if self._warn_user:
-            LOG.warning(self._api_key_error_msg)
-            self._warn_user = False
-
-    def _backoff(self, attempt: int) -> None:
-        backoff = self._backoff_base ** (attempt + 2)
-        LOG.trace(self._backoff_msg.format(backoff=backoff))
-        time.sleep(backoff)
-
-    def __call__(self, func: Callable[_P, _T]) -> Callable[_P, _T]:
-        @functools.wraps(func)
-        async def _wrapped(*args: _P.args, **kwargs: _P.kwargs) -> _T:
-            num_attempts = self._num_retries + 1  # +1 for the first attempt
-            for attempt in range(1, num_attempts + 1):
-                try:
-                    return await func(*args, **kwargs)
-
-                except RateLimitError:
-                    if attempt == num_attempts:
-                        raise
-                    self._log_rate_limit_error()
-
-                except APIError as e:
-                    if (e.http_status != 502) or (attempt == num_attempts):
-                        raise
-                except Exception as e:
-                    LOG.warning(e)
-                self._backoff(attempt)
-
-        return _wrapped
-
-
-def format_function_specs_as_typescript_ns(
-    functions: list[CompletionModelFunction],
-) -> str:
-    """Returns a function signature block in the format used by OpenAI internally:
-    https://community.openai.com/t/how-to-calculate-the-tokens-when-using-function-call/266573/18
-
-    For use with `count_tokens` to determine token usage of provided functions.
-
-    Example:
-    ```ts
-    namespace functions {
-
-    // Get the current weather in a given location
-    type get_current_weather = (_: {
-    // The city and state, e.g. San Francisco, CA
-    location: string,
-    unit?: "celsius" | "fahrenheit",
-    }) => any;
-
-    } // namespace functions
-    ```
-    """
-
-    return (
-        "namespace functions {\n\n"
-        + "\n\n".join(format_openai_function_for_prompt(f) for f in functions)
-        + "\n\n} // namespace functions"
-    )
-
-
-def format_openai_function_for_prompt(func: CompletionModelFunction) -> str:
-    """Returns the function formatted similarly to the way OpenAI does it internally:
-    https://community.openai.com/t/how-to-calculate-the-tokens-when-using-function-call/266573/18
-
-    Example:
-    ```ts
-    // Get the current weather in a given location
-    type get_current_weather = (_: {
-    // The city and state, e.g. San Francisco, CA
-    location: string,
-    unit?: "celsius" | "fahrenheit",
-    }) => any;
-    ```
-    """
-
-    def param_signature(name: str, spec: JSONSchema) -> str:
-        return (
-            f"// {spec.description}\n" if spec.description else ""
-        ) + f"{name}{'' if spec.required else '?'}: {spec.typescript_type},"
-
-    return "\n".join(
-        [
-            f"// {func.description}",
-            f"type {func.name} = (_ :{{",
-            *[param_signature(name, p) for name, p in func.parameters.items()],
-            "}) => any;",
-        ]
-    )
-
-
-def count_openai_functions_tokens(
-    functions: list[CompletionModelFunction], count_tokens: Callable[[str], int]
-) -> int:
-    """Returns the number of tokens taken up by a set of function definitions
-
-    Reference: https://community.openai.com/t/how-to-calculate-the-tokens-when-using-function-call/266573/18
-    """
-    return count_tokens(
-        f"# Tools\n\n## functions\n\n{format_function_specs_as_typescript_ns(functions)}"
-    )
-
-
-def _functions_compat_fix_kwargs(
-    functions: list[CompletionModelFunction],
-    completion_kwargs: dict,
-):
-    function_definitions = format_function_specs_as_typescript_ns(functions)
-    function_call_schema = JSONSchema(
-        type=JSONSchema.Type.OBJECT,
-        properties={
-            "name": JSONSchema(
-                description="The name of the function to call",
-                enum=[f.name for f in functions],
-                required=True,
-            ),
-            "arguments": JSONSchema(
-                description="The arguments for the function call",
-                type=JSONSchema.Type.OBJECT,
-                required=True,
-            ),
-        },
-    )
-    tool_calls_schema = JSONSchema(
-        type=JSONSchema.Type.ARRAY,
-        items=JSONSchema(
-            type=JSONSchema.Type.OBJECT,
-            properties={
-                "type": JSONSchema(
-                    type=JSONSchema.Type.STRING,
-                    enum=["function"],
-                ),
-                "function": function_call_schema,
-            },
-        ),
-    )
-
-    completion_kwargs["messages"] = [
-        OpenAIChatMessage.system(
-            "# tool usage instructions\n\n"
-            "Specify a '```tool_calls' block in your response,"
-            " with a valid JSON object that adheres to the following schema:\n\n"
-            f"{tool_calls_schema.to_dict()}\n\n"
-            "Specify any tools that you need to use through this JSON object.\n\n"
-            "Put the tool_calls block at the end of your response"
-            " and include its fences if it is not the only content.\n\n"
-            "## functions\n\n"
-            "For the function call itself, use one of the following"
-            f" functions:\n\n{function_definitions}"
-        ),
-    ]
-
-
-def _tool_calls_compat_extract_calls(response: str) -> list[AssistantToolCallDict]:
-    import json
-    import re
-
-    LOG.trace(f"Trying to extract tool calls from response:\n{response}")
-
-    if response[0] == "[":
-        tool_calls: list[AssistantToolCallDict] = json.loads(response)
-    else:
-        block = re.search(r"```(?:tool_calls)?\n(.*)\n```\s*$", response, re.DOTALL)
-        if not block:
-            raise ValueError("Could not find tool calls block in response")
-        tool_calls: list[AssistantToolCallDict] = json.loads(block.group(1))
-
-    for t in tool_calls:
-        t["function"]["arguments"] = str(t["function"]["arguments"])  # HACK
-    return tool_calls
