+++ AFAAS/core/agents/usercontext/models.py
@@ -12,7 +12,7 @@ if TYPE_CHECKING:
 class UserContextAgentSystems(BaseAgentSystems):
 
     ability_registry : str  ="AFAAS.core.ability.SimpleToolRegistry"
-    chat_model_provider : str  = "AFAAS.core.adapters.openai.OpenAIProvider"
+    chat_model_provider : str  = "AFAAS.core.adapters.openai.chatmodel.AFAASChatOpenAI"
     prompt_manager : str ="AFAAS.interfaces.agent.PromptManager"
     
     class Config(BaseAgentSystems.Config):
