+++ AFAAS/prompts/planner/strategies/__init__.py
@@ -18,7 +18,7 @@ class StrategiesConfiguration(PromptStrategiesConfiguration):
 
 class StrategiesSet:
     from AFAAS.interfaces.prompts.strategy import (
-        AbstractPromptStrategy, AbstractPromptStrategy)
+        AbstractPromptStrategy)
 
     @staticmethod
     def get_strategies() -> list[AbstractPromptStrategy]:
